Fixing merge issues

+++ Code/TomoEngine/IO/VTKWriterMacros.h
@@ -97,9 +97,9 @@
   fprintf(f, "LOOKUP_TABLE default\n"); \
   { \
     float t;\
-    for (uint16_t i = 0; i < ptr->N_z; ++i) {\
-     for (uint16_t j = 0; j < ptr->N_y; ++j) {\
-      for (uint16_t k = 0; k < ptr->N_x; k++) {\
+    for (int i = 0; i < ptr->N_z; ++i) {\
+     for (int j = 0; j < ptr->N_y; ++j) {\
+      for (int k = 0; k < ptr->N_x; k++) {\
           t = static_cast<float>(ptr->Object->d[i][k][j]);\
           MXA::Endian::FromSystemToBig::convert<float>(t); \
           fwrite(&t, sizeof(float), 1, f);\
